최종 엔티티, 피쳐 기준 변경
