+++ src/entities/posts/api/index.ts
@@ -1,3 +1,4 @@
 export * from './query';
 export * from './remote';
 export * from './query-keys';
+export * from './usePostsQuerykey';
