+++ src/entities/posts/api/remote.ts
@@ -1,13 +1,7 @@
 import type {
-  AddPostCommentResponse,
-  AddPostRequest,
-  AddPostResponse,
-  Post,
   PostCommentsResponse,
   PostsResponse,
   PostsTagsResponse,
-  UpdatePostCommentRequest,
-  UpdatePostCommentResponse,
   PostSortBy,
   SortOrder,
 } from '../model';
@@ -91,68 +85,4 @@ export const getPostComments = async (
   postId: number,
 ): Promise<PostCommentsResponse> => {
   return await remote(`/api/comments/post/${postId}`);
-};
-
-export const patchPostCommentLike = async (
-  commentId: number,
-  likes: number,
-) => {
-  return await remote(`/api/comments/${commentId}`, {
-    method: 'PATCH',
-    headers: { 'Content-Type': 'application/json' },
-    data: { likes },
-  });
-};
-
-export const deletePostComment = async (commentId: number) => {
-  return await remote(`/api/comments/${commentId}`, {
-    method: 'DELETE',
-  });
-};
-
-export const addPostComment = async (
-  postId: number,
-  userId: number,
-  body: string,
-): Promise<AddPostCommentResponse> => {
-  return await remote('/api/comments/add', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    data: { body, postId, userId },
-  });
-};
-
-export const updatePostComment = async (
-  commentId: number,
-  commentData: UpdatePostCommentRequest,
-): Promise<UpdatePostCommentResponse> => {
-  return await remote(`/api/comments/${commentId}`, {
-    method: 'PUT',
-    headers: { 'Content-Type': 'application/json' },
-    data: commentData,
-  });
-};
-
-export const updatePost = async (postId: number, post: Post): Promise<Post> => {
-  return await remote(`/api/posts/${postId}`, {
-    method: 'PUT',
-    headers: { 'Content-Type': 'application/json' },
-    data: post,
-  });
-};
-
-export const deletePost = async (postId: number) => {
-  return await remote(`/api/posts/${postId}`, {
-    method: 'DELETE',
-  });
-};
-
-export const addPost = async (
-  postData: AddPostRequest,
-): Promise<AddPostResponse> => {
-  return await remote('/api/posts/add', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    data: postData,
-  });
 };
