+++ src/entities/posts/model/lib/index.ts
@@ -1,2 +1,1 @@
-export * from './comment';
 export * from './post';
