+++ src/features/posts/api/index.ts
@@ -1,0 +1,2 @@
+export * from './remote';
+export * from './types';
