+++ src/features/posts/index.ts
@@ -1,2 +1,2 @@
 export * from './ui';
-export * from './model';
+export * from './model/useSearch';
