+++ src/features/posts/model/index.ts
@@ -1,6 +1,4 @@
-export * from './usePostDialog';
-export * from './useAddCommentDialog';
-export * from './useEditPostDialog';
-export * from './useAddPostDialog';
-export * from './useEditCommentDialog';
 export * from './useSearch';
+export * from './useMutation';
+export * from './mutate';
+export * from './useDialog';
