+++ src/features/users/index.ts
@@ -1,2 +1,1 @@
 export * from './ui';
-export * from './model';
