+++ src/features/users/ui/index.ts
@@ -1,1 +1,1 @@
-export { UserModal } from './UserModal';
+export { UserDialog, OpenUserDetailButton } from './UserDetail';
